Fix JIT
